Changed subjects and extracted mains

+++ mbta/treatment/representation/visitor/removemainvisitor.py
@@ -7,22 +7,23 @@ from mbta.treatment.representation.visitor.classvisitor import ClassVisitor
 class RemoveMainVisitor(ClassVisitor):
     def visit_java_file(self, source_file):
         source_path: Path = source_file.source_path
-        with source_path.open("r+") as file_reader_writer:
-            content = file_reader_writer.readlines()
-            file_reader_writer.seek(0, 0)
-            inject_imports = "import java.io.FileWriter;\nimport java.io.IOException;\n"
-            if inject_imports not in content:
-                file_reader_writer.write(inject_imports)
-            copy = True
-            for line in content:
-                if copy:
-                    if line.startswith("//TOFILL"):
-                        copy = False
-                    file_reader_writer.write(line)
-                else:
-                    file_reader_writer.write("\n}\n")
-                    break
-            file_reader_writer.truncate()
+        if "_MAIN" not in source_path.stem:
+            with source_path.open("r+") as file_reader_writer:
+                content = file_reader_writer.readlines()
+                file_reader_writer.seek(0, 0)
+                inject_imports = ["import java.io.FileWriter;\n", "import java.io.IOException;\n"]
+                if not inject_imports == content[:2]:
+                    file_reader_writer.write("".join(inject_imports))
+                copy = True
+                for line in content:
+                    if copy:
+                        if line.startswith("//TOFILL"):
+                            copy = False
+                        file_reader_writer.write(line)
+                    else:
+                        file_reader_writer.write("\n}\n")
+                        break
+                file_reader_writer.truncate()
 
     def visit_python_file(self, source_file):
         pass
